+++ CHANGELOG.rst
@@ -20,6 +20,8 @@ Added
 
 Changed
 -------
+- validate_{} functions for slots now return dictionaries of form {slot: value} 
+  instead of value
 
 Fixed
 -----